Strict Types requirements for PointUtils.h - GA_Offset does not implicitly cast to size_t.
Remove unneeded OPENVDB_HOUDINI_API as this hook isn't exported.

Houdini 16.0 uses version 330, not 150.

+++ openvdb_houdini/houdini/PointUtils.h
@@ -529,7 +529,7 @@ struct HoudiniReadAttribute
 
     size_t size() const
     {
-        return mOffsets ? mOffsets->size() : mAttribute.getIndexMap().indexSize();
+        return mOffsets ? mOffsets->size() : size_t(mAttribute.getIndexMap().indexSize());
     }
 
 private:
